agritech_backend/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-cors-headers==4.8.0
Pillow==10.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
requests==2.31.0
python-multipart==0.0.6
